chore: cleanup, async loaders added, duplicate fetches removed
