+++ src/components/Loader/Loader.tsx
@@ -1,0 +1,27 @@
+import React from 'react'
+import './Loader.scss'
+
+interface LoaderProps {
+  type: 'spinner' | 'ripple'
+}
+export const Loader: React.FC<LoaderProps> = ({
+  type
+}) => {
+  switch (type) {
+    case 'ripple': return (
+          <div className="mx-auto w-fit-content">
+
+          <div className="lds-ripple mx-auto">
+                <div />
+                <div />
+            </div>
+          </div>
+    )
+    default: return (
+            <div className="lds-ripple">
+                <div />
+                <div />
+            </div>
+    )
+  }
+}
